+++ app/src/main/java/com/example/dmitry/diplom_averin/rest/GraphicRest.java
@@ -12,18 +12,18 @@ import java.util.List;
 
 public class GraphicRest {
 
-    public GraphicRest(List<Pair<Double,Double>> list) {
+    public GraphicRest(List<Pair<Integer,Integer>> list) {
         points = list;
     }
 
     @SerializedName("points")
-    private List<Pair<Double,Double>> points;
+    private List<Pair<Integer,Integer>> points;
 
-    public List<Pair<Double, Double>> getPoints() {
+    public List<Pair<Integer, Integer>> getPoints() {
         return points;
     }
 
-    public void setPoints(List<Pair<Double,Double>> _points) {
+    public void setPoints(List<Pair<Integer,Integer>> _points) {
         points = _points;
     }
